db refactored

+++ src/server/routes/user.ts
@@ -1,16 +1,20 @@
 import express, { Response } from 'express';
-import User from '../models/User.ts';
+import User from '../../db/models/User.ts';
 import tokenService from '../services/token.service.ts';
 import { AuthedRequest, checkAuth } from '../middleware/auth.middleware.ts';
 import showError from '../../utils/console/showError.ts';
-import Operation from '../models/Operation.ts';
-import Category from '../models/Category.ts';
-import Account from '../models/Account.ts';
-import Goal from '../models/Goal.ts';
+import Operation from '../../db/models/Operation.ts';
+import Category from '../../db/models/Category.ts';
+import Account from '../../db/models/Account.ts';
+import Goal from '../../db/models/Goal.ts';
 import sendAuthError from '../../utils/errors/fromServerToClient/sendAuthError.ts';
 import sendForbidden from '../../utils/errors/fromServerToClient/sendForbidden.ts';
 import { sendNotFound } from '../../utils/errors/fromServerToClient/sendNotFound.ts';
 import serverError from '../../utils/errors/fromServerToClient/serverError.ts';
+import cryptService from '../services/crypt.service.ts';
+import sendCredentialsError from '../../utils/errors/fromServerToClient/sendCredentialsError.ts';
+import { cyanLog } from '../../utils/console/coloredLogs.ts';
+import showElement from '../../utils/console/showElement.ts';
 
 const router = express.Router({ mergeParams: true });
 
@@ -41,14 +45,26 @@ async function updateUser(req: AuthedRequest, res: Response) {
 async function removeUser(req: AuthedRequest, res: Response) {
   const thisPlace = 'user/remove';
   try {
+    cyanLog('user delete requested');
     if (!req.user) return sendAuthError(res, thisPlace);
+    cyanLog('auth check passed');
+
     const { id } = req.params;
-    const isPermitted = (id === req.user._id);
-    if (!isPermitted) return sendForbidden(res, thisPlace);
     const removingUser = await User.findById(id);
     if (!removingUser) return sendNotFound(res, "user", id);
+    cyanLog('user existence check passed');
+
+    if (!req?.headers?.authorization) return sendForbidden(res, thisPlace);
+    const password = req.headers.authorization.split(' ')[1];
+    const passwordIsOk = cryptService.compare(password, removingUser.password);
+    if (!passwordIsOk) sendCredentialsError(res);
+    cyanLog('password check passed')
+    const isPermitted = (id === req.user._id);
+    if (!isPermitted) return sendForbidden(res, thisPlace);
+    cyanLog('permition check passed');
 
     const result = await removingUser.deleteOne();
+    showElement(result, 'result');
     if (result.deletedCount) {
       await tokenService.removeTokens(removingUser._id);
       await removeUserData(req.user._id);
